definitivo/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
astroid==2.4.2
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.11.2
blinker==1.6.2
blis==0.7.9
catalogue==2.0.8
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
confection==0.0.4
cymem==2.0.7
es-core-news-md @ https://github.com/explosion/spacy-models/releases/download/es_core_news_md-3.5.0/es_core_news_md-3.5.0-py3-none-any.whl
Flask==2.3.2
Flask-Cors==4.0.0
frozenlist==1.3.3
gunicorn==21.2.0
idna==3.4
importlib-metadata==6.8.0
isort==5.6.2
itsdangerous==2.1.2
Jinja2==3.1.2
langcodes==3.3.0
lazy-object-proxy==1.4.3
MarkupSafe==2.1.2
mccabe==0.6.1
multidict==6.0.4
murmurhash==1.0.9
numpy==1.24.3
openai==0.27.7
packaging==23.1
pathy==0.10.1
preshed==3.0.8
pydantic==1.10.8
pylint==2.6.0
python-dotenv==1.0.0
requests==2.31.0
six==1.15.0
smart-open==6.3.0
soupsieve==2.4
spacy==3.5.3
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
thinc==8.1.10
toml==0.10.1
tqdm==4.65.0
typer==0.7.0
typing-extensions==4.6.2
urllib3==2.0.2
wasabi==1.1.1
Werkzeug==2.3.6
wrapt==1.12.1
yarl==1.9.2
zipp==3.16.2
